se completo el registro de usuario y registro controller satisfactoriamente

+++ public/extras/upload.php
@@ -6,9 +6,15 @@ if($_FILES["file"]["name"] != '')
  $ext = end($test);
  $name = rand(100,110) . '.' . $ext;
  $ubicacion = './tmp/'.$name;  
- 
- move_uploaded_file($_FILES["file"]["tmp_name"], $ubicacion);
+ if(file_exists($ubicacion)){
+    unlink($ubicacion);
+}
+
+if (!file_exists($ubicacion)) {
+    move_uploaded_file($_FILES["file"]["tmp_name"],$ubicacion);
+}
  $location='./extras/tmp/'.$name;
  echo '<img src="'.$location.'" height="350" width="500" class="img-thumbnail" />';
+ echo  '<input type="text" id="FOTO" name="FOTO" value="'.$location.'" style="display: none">';
 }
 ?>
